Finally got hold of the hardware. Made changes to fix the various issues with screen size, orientation, colour order, and mutexes never freeing. Also added a cut down version of the Fira Code font as a SPIFFS file for use in the UI.

+++ include/uielement.h
@@ -10,7 +10,7 @@
 #include "processdata.h"
 namespace euc {
 
-using ColourProvider = std::function<uint32_t(ProcessData*)>;
+using ColourProvider = std::function<uint16_t(ProcessData*)>;
 
 #define UI_SETUP(element, ...)                                          \
   public: static UiElement* Builder(uint8_t data[], size_t data_len) {  \
